feat: Add Game_of_Stones.kt
description:
Two players called P1 and P2 are playing a game with a starting number of stones. Player1 always plays first, and the two players move in alternating turns. The game's rules are as follows:
*In a single move, a player can remove either 2, 3 or 5 stones from the
game board.
*If a player is unable to make a move, that player loses the game.
Given the starting number of stones, find and print the name of the
winner.